backend/requirements.txt
# FastAPI và dependencies
fastapi==0.120.0
uvicorn[standard]==0.29.0
python-multipart==0.0.20

# SQLModel và Database
sqlmodel==0.0.27
SQLAlchemy==2.0.44
pymysql==1.1.0

# Pydantic
pydantic[email]==2.10.0

# Image Processing
Pillow>=10.0.0

# File Upload
aiofiles==23.2.1

# Environment Variables
python-dotenv==1.0.0

# Testing (optional)
pytest>=7.0.0
httpx==0.27.0